Address review comments
Add /proc directory to the initramfs structure, rather than creating it as part of the init script.
Make busybox install the links to the builtin utilities as part of the init script, so that commands don't have to be prepended with busybox (which seemed to only be required in some cases).
Modify the kernel make directive to use the number of processor cores available to the Github runner to speed up the build process.