Verify modules will compile for amd64
The user-mode linux architecture compile hides errors we may encounter when
actually compiling for release. Therefore, we need to ensure our modules
can at least be compiled against amd64.

Introduce 3 new steps:

1. make allnoconfig
Ensures that we're not introducing anything that would accidentally break
a build where nothing is set.

2. make defconfig
This is our control build.

3. make defconfig + config-integration
This tests that we can do our control + custom options. i.e. enable
modules we want to compile.

These three tests should be enough to satisfy Intel's testing-bot. We'll skip
allyesconfig since that takes a very long time to compile. Lastly,
enable "continue-on-error" so that we can see the results for all 3
builds at once and not individually.

Signed-off-by: Frederick Lawler <fred@cloudflare.com>